refactor(csharp): Cleanup C API
